ACCUMULO-2583 First stab at setting up the actual wire transfer to the replication peer
Changed a few things in the thrift idl to make a little more sense. Use the protobuf
Status message that we already have as the response object to the replication. This
allows the peer to not have to process the entire message at one time as an optimization (later).

Made a ReplicationClient in the same spirit as the MasterClient, but with a few notable
changes. Since we're operating outside of the local instance, we don't want to make
the same assertions about availability and blocking until we get a connection as is
the case in talking to the master. This pushes a bit of the logic into the replication
setup code, but this is desirable as we can better handle failure/retry logic as we intend.

Conflicts:
	core/src/main/java/org/apache/accumulo/core/client/impl/ConnectorImpl.java
	core/src/main/java/org/apache/accumulo/core/client/impl/MasterClient.java