ACCUMULO-2575 Fix bug in GC where files were being prematurely deleted and add tests to catch
After realizing the deadlock potential when not having tservers write the initial
replication entry to the metadata table, the garbage collection code was not updated
for the shift in data layout. On a clean system, when first beginning replication,
the replication table will not yet exist, but replication entries will exist in the
metadata table.

The GC code still assuming that the lack of a replication table implied that
there was no data pending replication and short-circuited early. We need to check
the metadata table always, and, when present, also the replication table.