ACCUMULO-2575 Update RemoveCompleteReplicationRecords to remove closed, fully replicated entries from both tables.
Because we're lazily passing the state between the metadata and replication tables,
we must scan and submit deletes for each of them separately. There is a race condition
where we might delete an element from metadata only to have the StatusMaker re-make
a record in the replication, but this is acceptable as future invocations of this
class will clean it up for us.