ACCUMULO-2819 Initial pass at a new WorkAssigner that will only replicate the oldest file per table per peer
By tracking the order in which WALs were MinC'ed, we can use this order to determine
the practical order for which they should be replayed on each table of each peer. This
aims to remove the potential problem that the DWQWorkAssigner faced in that WAL2 might be
replicated before WAL1 (where WAL1 was written before WAL2 locally).