ACCUMULO-2574 Change closedTime into createdTime
It's rather impossible to actually konw the exact point in time when
a WAL is no longer in use by any tablets. We can know when a tablet is
done with a WAL, but many tablets may be referencing it which convolutes
things.

Instead, we can get the same intent (global ordering of WALs for a tserver)
by recording the time in which the WAL was first created as this is something
that only occurs once in the lifetime of a file.

Conflicts:
	core/src/main/java/org/apache/accumulo/core/replication/StatusUtil.java