ACCUMULO-378 Removing experimental annotation.
In talking to [~ctubbsii] on the subject, he stated that the intent of Experimental
was for features that are incomplete or not expected to fully work. Replication is
not one of those features (it is tested and expected to work with known functionality),
therefore I'm removing the experimental annotation.

Conflicts:
	core/src/main/java/org/apache/accumulo/core/conf/Property.java