ACCUMULO-2925 Need to preserve replicationSource on the Mutation
The replicationSource on the Mutation is the information which prevents cycles
in the replication graph from infinitely replicating information. Each replicationSource
on a Mutation is the `replication.name` for a system from which that Mutation came.

We can later use this set to determine if we need to replicate this Mutation
to a given peer by observing if the `replication.name` of our peer already
exists in the replicationSources.