ACCUMULO-3077 Set StatusCombiner on accumulo.metadata during initialize.
Before mutations are written to the metadata table to
track replication, the configuration for the metadata
table is checked to ensure that the Combiner is set
(to make sure that the state is appropriately managed).

Because we cache ZK data and update it via Watchers, it's
possible that a tserver compacts some data after it was written
before seeing that the Combiner was set, effectively destroying
the older versions (VersioningIterator).

Setting the Combiner at init for the metadata table
does a better job at eliminiating this possibility.

Conflicts:
	server/tserver/src/main/java/org/apache/accumulo/tserver/tablet/DatafileManager.java