ACCUMULO-3233 retry to start the minicluster when ZK fails to start
There's a race condition between choosing a random port and when
zookeeper actually binds to that port in which another process
could bind to the port first. This results in accumulo initialization
to fail because ZK died.

This is a very infrequent occurrence, so we can just retry a few times
and that should reduce the frequency in which we see it happen.

Conflicts:
	test/src/test/java/org/apache/accumulo/test/functional/SimpleMacIT.java