ACCUMULO-3215 Ensure tablet directories are absolute, not relative
Move 'c-' into a constant, make the computed srv:dir in ImportTable
an absolute directory (choosing a volume as per policy for each tablet),
and add a test to make sure the path is constructed correctly.

ImportExportIT was extended to check that srv:dir, in addition to file:,
is an absolute path.

Conflicts:
	server/master/pom.xml