ACCUMULO-3264 Only reset logging configuration when we're certain that we have configuration files
The original change broke logging for all of the ITs because
the logging configuration was reset without having any means
to add back the original log4j configuration from maven.

This shouldn't have been a problem because MAC should really
be setting up logging in the same way that a real cluster does
instead of faking it.