ACCUMULO-3279 Remove version argument from recursiveDelete
`recursiveDelete` is much like a big hammer: get rid of
all of the nodes beneath this parent. As such, it doesn't
make sense to pass in a version, nor does it make sense to
provide a specific version to delete. We can pass in a version
of '-1' which should eliminate the need for any special handling
around BadVersionExceptions.

Conflicts:
	fate/src/main/java/org/apache/accumulo/fate/zookeeper/IZooReaderWriter.java
	fate/src/main/java/org/apache/accumulo/fate/zookeeper/ZooReaderWriter.java
	fate/src/main/java/org/apache/accumulo/fate/zookeeper/ZooUtil.java