BUG-36879 Hack around netty-all dependencies not coming in transitively
hadoop-hdfs needs netty which we pull in via
hadoop-minicluster and directly. For some reason, with
the HWX builds, we don't get it. In test environments,
we miss this dependency and the tests that use an HDFS
minicluster fail on a classnotfoundexception.

This adds a test dependency to deal with this, ignoring
what should be a root cause fix in Hadoop.