BUG-124710 ACCUMULO-4421 Ensure that TraceServer launches the Kerberos ticket renewal thread
The previous method that TraceServer was using on SecurityUtil
only performed the login and was relying on incorrect functionality
in Hadoop's UGI to launch a renewal thread. This logic is incorrect.
Refactored SecurityUtil a little to prevent other callers from
making the same mistake in the future.

Signed-off-by: Sean Busbey <busbey@cloudera.com>
(cherry picked from commit 72f77994ff06f8ddcfb291d12512d956c23a2823)