fixes #767 correct allocation issue in native maps (#769)
This fixes a bug that was found when building the native maps against
Fedora 29.  The cause of the bug was that an empty C++ map seemed to
allocate a small amount of memory, which was a new behavior.  This new
behavior violated an assumption made by the custom alloctor used by the
native map code.  The code was restructured to avoid this issue.

Change-Id: I150d0d37044b5c9b3dff5e6d16e008c4df269fc5